More classes!
Fixed nested classes in Product class file. Now every sub-class has it's
own little file!
